requirements.txt
pillow==6.0.0
numpy==1.16.4
pandas==0.24.2
scikit-image==0.15.0
matplotlib==3.1.0
Flask==1.0.2
Werkzeug<0.13.0,>=0.12.0
flask-restplus
Flask-Cors
tensorflow==1.13.1
keras==2.2.4